+++ source/B2B.Transactions.Api/Middleware/Authentication/MarketActors/MarketActorAuthenticatorMiddleware.cs
@@ -14,7 +14,7 @@
 
 using System;
 using System.Threading.Tasks;
-using B2B.Transactions.Authentication;
+using B2B.Transactions.Configuration.Authentication;
 using B2B.Transactions.Infrastructure.Authentication;
 using B2B.Transactions.Infrastructure.Serialization;
 using Microsoft.Azure.Functions.Worker;
