+++ source/B2B.Transactions.Infrastructure/Authentication/MarketActorAuthenticator.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Linq;
 using System.Security.Claims;
-using B2B.Transactions.Authentication;
+using B2B.Transactions.Configuration.Authentication;
 
 namespace B2B.Transactions.Infrastructure.Authentication
 {
