+++ source/B2B.Transactions.Infrastructure/Configuration/CompositionRoot.cs
@@ -17,8 +17,8 @@ using Azure.Messaging.ServiceBus;
 using B2B.CimMessageAdapter;
 using B2B.CimMessageAdapter.Messages;
 using B2B.CimMessageAdapter.Transactions;
-using B2B.Transactions.Authentication;
 using B2B.Transactions.Configuration;
+using B2B.Transactions.Configuration.Authentication;
 using B2B.Transactions.DataAccess;
 using B2B.Transactions.IncomingMessages;
 using B2B.Transactions.Infrastructure.Authentication;
