+++ source/B2B.Transactions.IntegrationTests/Infrastructure/Authentication/MarketActors/MarketActorAuthenticatorTests.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Collections.Generic;
 using System.Security.Claims;
-using B2B.Transactions.Authentication;
+using B2B.Transactions.Configuration.Authentication;
 using B2B.Transactions.Infrastructure.Authentication;
 using Xunit;
 using Xunit.Categories;
